% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{create_helloworld_api}
\alias{create_helloworld_api}
\title{Initialize Hello World API using RestRserve}
\usage{
create_helloworld_api()
}
\value{
A configured RestRserve application object
}
\description{
Creates and configures a RestRserve application for the Hello World functionality
}
\details{
Create Hello World API Application
}
\examples{
\dontrun{
app <- create_helloworld_api()
backend <- BackendRserve$new()
backend$start(app, http_port = 8000)
}
}
