% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helloWorld.R
\name{getAvailableLanguages}
\alias{getAvailableLanguages}
\title{List Available Languages}
\usage{
getAvailableLanguages()
}
\value{
A character vector of available languages
}
\description{
Returns a list of all available languages in the Hello_World dataset
}
\details{
Get Available Languages
}
