\name{packageSkeleton-package}
\alias{packageSkeleton-package}
\docType{package}
\title{A GitHub repository featuring a starter template for R package development, paired with a Quarto-powered website.
}
\description{
The packageSkeleton R package provides a comprehensive template to jumpstart the development of an R package, complete with a structured framework for coding, documentation, and version control. It simplifies the process of setting up an R package repository on GitHub, offering built-in integration with Quarto for generating a professional, customizable website to showcase package documentation and vignettes. This template is designed to help both new and experienced R developers streamline package creation while ensuring best practices in documentation and collaboration. With minimal setup, users can focus on building and sharing valuable tools within the R ecosystem.
}

\details{
\tabular{ll}{
Package: \tab packageSkeleton\cr
Type: \tab Package\cr
Version: \tab 0.0-0.1\cr
Date: \tab 2024-10-14\cr
License: \tab MIT\cr
LazyLoad: \tab yes\cr
}
}

\author{Damian W. Betebenner \email{DBetebenner@nciea.org}}

\keyword{package}
