% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{run_helloworld_api}
\alias{run_helloworld_api}
\title{Start Hello World API Server}
\usage{
run_helloworld_api(port = 8000, host = "127.0.0.1")
}
\arguments{
\item{port}{The port number to listen on (default: 8000)}

\item{host}{The host address to listen on (default: "127.0.0.1")}
}
\value{
None
}
\description{
Starts the Hello World API server using RestRserve
}
\details{
Run Hello World API
}
\examples{
\dontrun{
run_helloworld_api(port = 8000)
}
}
