% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strTail.R
\name{strTail}
\alias{strTail}
\title{Returns the Last \code{n} Characters of a String, with Support for Negative Values}
\usage{
strTail(string, n_char = 1)
}
\arguments{
\item{string}{Character string or vector of character strings.}

\item{n_char}{Number of characters to take from the tail (starting at the last character) of the
string (or vector of strings). A positive value returns the last \code{n_char} characters,
while a negative value removes the first \code{abs(n_char)} characters from the start. Defaults to 1.}
}
\value{
The function returns a character string or a vector of character strings, depending on the input.
The output consists of the last \code{n_char} characters from each string in the input. If \code{n_char}
is negative, the result excludes the first \code{abs(n_char)} characters from each string.
}
\description{
The \code{strTail} function extracts the last \code{n} characters of a provided string.
If \code{n_char} is positive, it returns the last \code{n_char} characters; if negative,
it returns all but the first \code{abs(n_char)} characters. The function handles input validation,
ensuring that \code{n_char} is within the bounds of the string length. It’s useful for quickly
trimming strings to desired lengths.
}
\details{
The \code{strTail} function is commonly used to extract the last few characters from a
character string or each string in a vector of character strings. This is helpful when you need to
retrieve suffixes or the end portion of strings. The function supports both positive and negative values
for \code{n_char}. For positive values, the function returns the last \code{n_char} characters, while
for negative values, it excludes the first \code{abs(n_char)} characters from the start of each string.
}
\examples{
# Example 1: Last two characters of each state name
strTail(state.name, 2)

# Example 2: Last three characters of a single string
strTail("California", 3)

# Example 3: Remove the first two characters from each string in a vector
strTail(c("apple", "banana", "cherry"), -2)

# Example 4: Get the last character of each month name
strTail(month.name, 1)

# Example 5: Handle a vector of different length strings
strTail(c("wolf", "elephant", "giraffe"), 4)

}
\author{
Damian W. Betebenner \email{dbetebenner@nciea.org}
}
\keyword{misc}
\keyword{models}
