% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSkeleton-package.R
\docType{package}
\name{packageSkeleton-package}
\alias{packageSkeleton}
\alias{packageSkeleton-package}
\title{A GitHub repository featuring a starter template for R package development, paired with a Quarto-powered website.}
\description{
packageSkeleton: A GitHub Repository Template for R Package Development with Quarto Integration
}
\details{
The \code{packageSkeleton} R package provides a comprehensive template to jumpstart the development of an R package,
complete with a structured framework for coding, documentation, and version control. It simplifies the process of
setting up an R package repository on GitHub, offering built-in integration with Quarto for generating a professional,
customizable website to showcase package documentation and vignettes. This template is designed to help both new
and experienced R developers streamline package creation while ensuring best practices in documentation and collaboration.
With minimal setup, users can focus on building and sharing valuable tools within the R ecosystem.

\tabular{ll}{
Package: \tab packageSkeleton \cr
Type: \tab Package \cr
Version: \tab 0.0-0.8 \cr
Date: \tab 2025-6-11 \cr
License: \tab MIT \cr
LazyLoad: \tab yes \cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://centerforassessment.github.io/packageSkeleton/}
  \item Report bugs at \url{https://github.com/centerforassessment/packageSkeleton/issues/}
}

}
\author{
\strong{Maintainer}: Damian W. Betebenner \email{dbetebenner@nciea.org} (\href{https://orcid.org/0000-0003-0476-5599}{ORCID})

Authors:
\itemize{
  \item Adam R. Van Iwaarden \email{avaniwaarden@nciea.org}
}

}
\keyword{package}
